all: handle 'V' unit for electrification in data only
Front does not add 'V' as unit anymore (but some hard-coded front
processing is not tackled and just migrated).
The responsibility for unit is on the data side (rolling stock,
catenaries, electrical profiles).

Hand-tested (cards, edit and curves on rolling-stock, legend and
catenaries on map, traction on study's speed-space graph):
- OK to load a dump, migrate and use
- OK to load a dump and use without migration (front crashes on speed-space
  graph as expected for traction)